🎨 PhoneCertificateView 명칭 변경
