+++ TiquiTaca_iOS/Views/VerificationNumberCheck/VerificationNumberCheckView.swift
@@ -1,15 +1,16 @@
 //
-//  PhoneCertificateView.swift
-//  SwiftUIPractice
+//  VerificationNumberCheckView.swift
+//  TiquiTaca_iOS
 //
-//  Created by 송하경 on 2022/04/17.
+//  Created by 강민석 on 2022/05/01.
 //
+
 import ComposableArchitecture
 import SwiftUI
 
-struct PhoneCertificateView: View {
-  let store: Store<PhoneCertificateState, PhoneCertificateAction>
-
+struct VerificationNumberCheckView: View {
+  let store: Store<VerificationNumberCheckState, VerificationNumberCheckAction>
+  
   var body: some View {
     WithViewStore(store) { viewStore in
       NavigationView {
@@ -23,21 +24,21 @@ struct PhoneCertificateView: View {
             OTPFieldView(
               store: store.scope(
                 state: \.otpFieldState,
-                action: PhoneCertificateAction.otpFieldAction
+                action: VerificationNumberCheckAction.otpFieldAction
               )
             )
           }
           
           Spacer()
-//          Button {
-//            viewStore.send(.certificationButtonTapped)
-//          }label: {
-//            Text("인증하기")
-//          }
-//          .disabled(checkStates())
-//          .opacity(checkStates() ? 0.4 : 1)
-//          .buttonStyle(NormalButtonStyle())
-//          .padding(20)
+          //          Button {
+          //            viewStore.send(.certificationButtonTapped)
+          //          }label: {
+          //            Text("인증하기")
+          //          }
+          //          .disabled(checkStates())
+          //          .opacity(checkStates() ? 0.4 : 1)
+          //          .buttonStyle(NormalButtonStyle())
+          //          .padding(20)
         }
       }
       .navigationBarHidden(true)
@@ -47,11 +48,11 @@ struct PhoneCertificateView: View {
 
 struct PhoneCertificateView_Previews: PreviewProvider {
   static var previews: some View {
-    PhoneCertificateView(
+    VerificationNumberCheckView(
       store: .init(
         initialState: .init(),
         reducer: phoneCertificateReducer,
-        environment: PhoneCertificateEnvironment(
+        environment: VerificationNumberCheckEnvironment(
           authService: .init(),
           mainQueue: .main
         )
